Rename modules update and async
