+++ bioconda_utils/hosters.py
@@ -33,7 +33,7 @@ from urllib.parse import urljoin
 import regex as re
 
 if TYPE_CHECKING:
-    from .async import AsyncRequests
+    from .aiopipe import AsyncRequests
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
